google: move openlineage imports inside methods
